[NEW] tracker field types: add a multiple selection list to specify which items should be listed to the users when editing a tracker item of the concerned field types.
This has been done for those field types :
	* user selector,
	* group selector,
	* country selector

It allows, for example, to list only developpers in a list of users in charge of bugs in a bug tracker, or to list only a few countries with their flags.

It is possible to use this for every field type that generates a drop-down list of items (categories, ...).

Choosing no item at all will list all items to the user, for compatibility with old versions, because it's easier to use this way and because nobody should need an empty list.