[FIX] many features: new TikiWiki version of parse_str, that:  - uses a workaround for a bug in PHP 5.2.0  - Handle the value of magic_quotes_gpc to stripslashes when needed (as already done for GET/POST/... in tiki-setup_base.php)
This fixes bugs in multiple places where the buggy parse_str was used (for example in smarty button, which caused problems when editing/renaming/... a wiki page that includes a ' in its name. Same problem occured, for example, when changing language of such a wiki page, ...).

Please now use TikiLib::parse_str() instead of parse_str().