[pbp/r20176][MOD] Add a pref to allow both filegals and the old wiki_up methods for adding images.
This is to allow quick uploads for "users in a hurry" (aren't they always?) but still enable the building of a library of images in an orderely fashion (e.g. using batch upload from img/wiki_up later).

* NB: pbp = Partial Back-port
In trunk this pref is settable in admin/wiki too, this MOD changes only the existing pref in admin/textarea